Fix CI issues: add test type context, extract validate-metrics, remove unused import
