+++ src/application/summary-logs/validate-metrics.js
@@ -1,0 +1,84 @@
+import { summaryLogMetrics } from '#common/helpers/metrics/summary-logs.js'
+import { ROW_OUTCOME } from '#domain/summary-logs/table-schemas/validation-pipeline.js'
+
+/** @import {ValidatedWasteRecord} from '#application/waste-records/transform-from-summary-log.js' */
+/** @import {ProcessingType} from '#domain/summary-logs/meta-fields.js' */
+
+/**
+ * Records validation issue metrics grouped by severity × category.
+ *
+ * @param {import('./validate-issue-logging.js').ValidationIssuesCollector} issues
+ * @param {string} processingType - The processing type for the metric dimension
+ */
+export const recordValidationIssueMetrics = async (issues, processingType) => {
+  const allIssues = issues.getAllIssues()
+  if (allIssues.length === 0) {
+    return
+  }
+
+  // Count issues by severity × category
+  const counts = new Map()
+  for (const issue of allIssues) {
+    const key = `${issue.severity}:${issue.category}`
+    counts.set(key, (counts.get(key) || 0) + 1)
+  }
+
+  // Record metrics for each combination
+  for (const [key, count] of counts) {
+    const [severity, category] = key.split(':')
+    await summaryLogMetrics.recordValidationIssues(
+      {
+        severity:
+          /** @type {import('#common/helpers/metrics/summary-logs.js').ValidationSeverity} */ (
+            severity
+          ),
+        category:
+          /** @type {import('#common/helpers/metrics/summary-logs.js').ValidationCategory} */ (
+            category
+          ),
+        processingType: /** @type {ProcessingType} */ (processingType)
+      },
+      count
+    )
+  }
+}
+
+/**
+ * Records row outcome metrics grouped by outcome.
+ *
+ * @param {ValidatedWasteRecord[] | null} wasteRecords - Waste records with outcomes
+ * @param {string} processingType - The processing type for the metric dimension
+ */
+export const recordRowOutcomeMetrics = async (wasteRecords, processingType) => {
+  if (!wasteRecords || wasteRecords.length === 0) {
+    return
+  }
+
+  // Count by outcome
+  const counts = {
+    [ROW_OUTCOME.INCLUDED]: 0,
+    [ROW_OUTCOME.EXCLUDED]: 0,
+    [ROW_OUTCOME.REJECTED]: 0,
+    [ROW_OUTCOME.IGNORED]: 0
+  }
+
+  for (const { outcome } of wasteRecords) {
+    counts[outcome]++
+  }
+
+  // Record metrics for each outcome with non-zero count
+  for (const [outcome, count] of Object.entries(counts)) {
+    if (count > 0) {
+      await summaryLogMetrics.recordRowOutcome(
+        {
+          outcome:
+            /** @type {import('#common/helpers/metrics/summary-logs.js').RowOutcome} */ (
+              outcome
+            ),
+          processingType: /** @type {ProcessingType} */ (processingType)
+        },
+        count
+      )
+    }
+  }
+}
